+++ todo-app/src/App.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
-import './App.css'; // Styling included here for presentable UI
-import todobgImage from './todobg.jpg'; // Adjust the name as necessar
-
+import './App.css'; 
+import TodoList from './TodoList';
+import InputForm from './InputForm';
 
 function TodoApp() {
   const [todos, setTodos] = useState([]);
@@ -42,74 +42,19 @@ function TodoApp() {
   return (
     <div className="todo-app">
       <h1>Todo List</h1>
-      <div className="input-container">
-        <input
-          type="text"
-          value={newTodo}
-          onChange={(e) => setNewTodo(e.target.value)}
-          placeholder="Enter a new todo"
-        />
-        <button onClick={addTodo}>Add Todo</button>
-      </div>
-
-      {todos.length === 0 ? (
-        <p>No todos available. Add a todo to get started!</p>
-      ) : (
-        <ul>
-          {todos.map((todo, index) => (
-            <TodoItem
-              key={index}
-              todo={todo}
-              index={index}
-              onComplete={handleComplete}
-              onEdit={handleEdit}
-              onSave={handleSave}
-              onRemove={handleRemove}
-            />
-          ))}
-        </ul>
-      )}
+      <InputForm 
+        newTodo={newTodo}
+        setNewTodo={setNewTodo}
+        addTodo={addTodo}
+      />
+      <TodoList
+        todos={todos}
+        handleComplete={handleComplete}
+        handleEdit={handleEdit}
+        handleSave={handleSave}
+        handleRemove={handleRemove}
+      />
     </div>
-  );
-}
-
-function TodoItem({ todo, index, onComplete, onEdit, onSave, onRemove }) {
-  const [editText, setEditText] = useState(todo.text);
-
-  return (
-    <li className={`todo-item ${todo.isComplete ? 'completed' : ''}`}>
-      {todo.isEditing ? (
-        <>
-          <input
-            type="text"
-            value={editText}
-            onChange={(e) => setEditText(e.target.value)}
-          />
-          <button onClick={() => onSave(index, editText)}>Save</button>
-        </>
-      ) : (
-        <>
-          <span>{todo.text}</span>
-          <button
-            onClick={() => onComplete(index)}
-          >
-            {todo.isComplete ? 'Undo' : 'Complete'}
-          </button>
-          <button
-            onClick={() => onEdit(index)}
-            disabled={todo.isComplete || todo.isEditing}
-          >
-            Edit
-          </button>
-          <button
-            onClick={() => onRemove(index)}
-            disabled={todo.isComplete || todo.isEditing}
-          >
-            Remove
-          </button>
-        </>
-      )}
-    </li>
   );
 }
